Centralize Critters Parser
