+++ src/templates/actor/tabs/InventoryTab.html
@@ -72,6 +72,6 @@
                 {{/if}}
             {{/each}}
         </div>
-        </div>
-        {{/'systems/shadowrun5e/dist/templates/common/TabWrapper.html'}}
+    </div>
+    {{/'systems/shadowrun5e/dist/templates/common/TabWrapper.html'}}
