+++ src/templates/apps/import.html
@@ -20,7 +20,8 @@
             Spells<input type="checkbox" class="spells" checked />
             Contacts<input type="checkbox" class="contacts" checked />
             Lifestyles<input type="checkbox" class="lifestyles" checked />
-            Vehicles<input type="checkbox" class="vehicles" />
+            Vehicles<input type="checkbox" class="vehicles" checked />
+            Critters<input type="checkbox" class="critters" checked />
         </div>
         <textarea class="chummer-text" rows="20" cols="50"> </textarea>
         <div class="flexrow">
